Add a userenv to the callback.pl background script
We were not previously setting a userenv in this background script and
as such the call to Koha::Account::pay was failing due to defaulting to
'0' for manager_id when no environment is found.

This patch adds an environment similar to that of an OPAC script with
the patrons details.

In more recent versions of Koha the default to '0' should have been
replaced by an 'undef' which would render this patch no longer required.